python/requirements.txt
websocket-client==0.55.0
